fixed bug in reduction code
typed arrays naturally know the type size and offset properly with a[i]
but void* doesn't know that so we need to use pointer arithmetic to give
the right offsets.

alternatively, we could explicitly case but that would lead to a lot
more code because C does not provide nice generic programming features
and i refuse to depend on C++ for this.